Lots of little improvements
