+++ mac/geral/templatetags/tags.py
@@ -10,10 +10,3 @@ def active(request, pattern):
     if re.search(pattern, request.path):
         return "active"
     return ""
-
-
-@register.simple_tag
-def display_none(request, pattern):
-    if re.search(pattern, request.path):
-        return "display:block"
-    return "display:none"
