+++ mac/publicacoes/urls.py
@@ -3,8 +3,7 @@ from django.conf.urls import url
 from .views import *
 
 urlpatterns = [
-    url(r"^$", index),
-    url(r"^catalogos/(?P<select>\w+)/$", catalogos),
-    url(r"^imprensa/(?P<select>\w+)/$", imprensa),
-    url(r"^critica/(?P<select>\w+)/$", critica),
+    url(r"^catalogos/(?P<select>\w+)/$", catalogos, name="catalogs"),
+    url(r"^imprensa/(?P<select>\w+)/$", imprensa, name="press"),
+    url(r"^critica/(?P<select>\w+)/$", critica, name="critics"),
 ]
