+++ mac/settings/base.py
@@ -54,10 +54,7 @@ STATICFILES_DIRS = [PROJECT_ROOT / "static"]
 STATIC_ROOT = BASE_DIR / "static"
 STATIC_URL = "/static/"
 
-# Make this unique, and don't share it with anybody.
-SECRET_KEY = os.getenv(
-    "SECRET_KEY", "w-ezr04of=m-vsp0om4b-dq7lkqqdf39gucpvs@gmi^@@q8hye"
-)
+SECRET_KEY = os.getenv("SECRET_KEY", "dev-secret")
 
 MIDDLEWARE = [
     "django.middleware.common.CommonMiddleware",
@@ -104,7 +101,6 @@ INSTALLED_APPS = (
     "mac.exposicoes",
     "mac.artistas",
     "mac.publicacoes",
-    "mac.contactos",
     "mac.geral",
     "storages",
 )
