+++ mac/templates/banner.html
@@ -1,0 +1,10 @@
+{% load static %}
+<div class="banner">
+    {% block tela %}
+    <div id="tela" class="banner__container">
+        {% for tela in telas %}
+            <img class="banner__img" src="{% get_media_prefix %}{{ tela.image }}" />
+        {% endfor %}
+    </div>
+    {% endblock %}
+</div>
